Fix max edit distance to ensure interface similarity between our aligner and 'bwa aln'.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2043 348d0f76-0448-11de-a6fe-93d51630548a

+++ c/bwa/bwa_gateway.h
@@ -32,7 +32,7 @@ class BWA {
   ~BWA();
 
   // Parameterize the aligner.
-  void set_max_edit_distance(int edit_distance);
+  void set_max_edit_distance(float edit_distance);
   void set_max_gap_opens(int max_gap_opens);
   void set_max_gap_extensions(int max_gap_extensions);
   void set_disallow_indel_within_range(int indel_range);
